% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{briere2_tpc}
\alias{briere2_tpc}
\title{Brière-2 thermal performance curve (TPC)}
\usage{
briere2_tpc(temp, ctmin, ctmax, a, b, scale = FALSE)
}
\arguments{
\item{temp}{Numeric vector of temperatures}

\item{ctmin}{Single numeric for parameter \code{ctmin}.}

\item{ctmax}{Single numeric for parameter \code{ctmax}.}

\item{a}{Single numeric for parameter \code{a}.}

\item{b}{Single numeric for parameter \code{b}.}

\item{scale}{Single logical for whether to scale to make max value 1.
Defaults to \code{FALSE}.}
}
\value{
A numeric vector for measure of performance for each in \code{temp}
}
\description{
Note that this TPC does not perform well when ctmin is positive, but
there are negative temperatures.
For example, if you run \code{briere2_tpc(c(-10, 0, 10), ctmin = 5, ctmax = 30, a = 1, b = 0.2, TRUE)}, you'll see that a temperature of \code{-10}
gives a greater performance value than \code{10}.
}
\examples{
briere2_tpc(temp = seq(10, 35, 5), ctmin = 8, ctmax = 38, a = 1, b = 0.2)

}
