% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{briere2_tpc_deriv}
\alias{briere2_tpc_deriv}
\title{Derivative of Brière-2 thermal performance curve (TPC) with respect to time}
\usage{
briere2_tpc_deriv(temp, ctmin, ctmax, a, b)
}
\arguments{
\item{temp}{Numeric vector of temperatures}

\item{ctmin}{Single numeric for parameter ctmin}

\item{ctmax}{Single numeric for parameter ctmax}

\item{a}{Single numeric for parameter a}

\item{b}{Single numeric for parameter b}
}
\value{
A numeric vector for first derivative of measures of
performance for each in \code{temp}
}
\description{
Derivative of Brière-2 thermal performance curve (TPC) with respect to time
}
