% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{briere2_tpc_Topt}
\alias{briere2_tpc_Topt}
\title{Thermal optimum for Brière-2 thermal performance curve (TPC)}
\usage{
briere2_tpc_Topt(ctmin, ctmax, b)
}
\arguments{
\item{ctmin}{Numeric vector for parameter \code{ctmin}.}

\item{ctmax}{Numeric vector for parameter \code{ctmax}.}

\item{b}{Numeric vector for parameter \code{b}.}
}
\value{
A numeric vector of optimum temperatures.
}
\description{
Note that parameter \code{a} is not necessary for this calculation.
}
\examples{
briere2_tpc_Topt(ctmin = 8, ctmax = 38, b = 0.2)

}
