% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting.R
\name{deriv_design_temps}
\alias{deriv_design_temps}
\title{Design experimental temperatures}
\usage{
deriv_design_temps(
  n_temps,
  z0,
  z1,
  ctmin,
  ctmax,
  a,
  b,
  temp_buffer = 5,
  temp_by = 0.1
)
}
\arguments{
\item{n_temps}{Single integer for the number of temperatures to sample from.
Must be >= 2.}

\item{z0}{Single numeric for the degree to which sampled temperatures
are derivative based versus uniform at the lowest possible
sampled temperature. Values of \code{0} indicate uniform, and values of \code{1}
indicate totally derivative based.
Must be >= 0.}

\item{z1}{Single numeric for the degree to which sampled temperatures
are derivative based versus uniform at the highest possible
sampled temperature. Values of \code{0} indicate uniform, and values of \code{1}
indicate totally derivative based.
Must be >= 0.}

\item{ctmin}{Single numeric for parameter \code{ctmin}.}

\item{ctmax}{Single numeric for parameter \code{ctmax}. Must be > \code{ctmin}.}

\item{a}{Single numeric for parameter \code{a}. Must be > 0.}

\item{b}{Single numeric for parameter \code{b}. Must be > 0.}

\item{temp_buffer}{Single numeric for the "buffer" around \code{ctmin} and \code{ctmax}
to have for the possible sampling temperatures.
More specifically, the minimum possible sampling temperature is
\code{ctmin - temp_buffer}, and the maximum is \code{ctmax + temp_buffer}
(both are rounded to have the same number of decimal places as
\code{temp_by}).
Must be >= 0.
Must not have more decimal places than those in \code{temp_by}.
Defaults to \code{5}.}

\item{temp_by}{Single numeric for the increment to generate possible
sampling temperatures by.
Specifically, possible temperatures will be
\code{seq(ctmin - temp_buffer, ctmax + temp_buffer, temp_by)}.
This number should reflect the precision of the equipment used to
generate temperatures.
Must be > 1e-6. Defaults to \code{0.1}.}
}
\value{
A vector of temperatures at which to sample.
}
\description{
Design experimental temperatures
}
