% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{make_temps}
\alias{make_temps}
\title{Make vector of temperatures for input of parameters}
\usage{
make_temps(params, temp_min, temp_max, n_temps)
}
\arguments{
\item{temp_min}{Single numeric for the minimum possible temperature allowed.}

\item{temp_max}{Single numeric for the maximum possible temperature allowed.}

\item{n_temps}{Single integer for the number of temperatures to sample.}

\item{temp}{Numeric vector of parameters. The first must be the
logit-transformed proportion the first temperature to be sampled is
from \code{temp_min} to \code{(temp_min + temp_max) / 2}.
The second is the same for the proportion the last temperature to
be sampled is from the midpoint to \code{temp_max}.
The third and fourth (if present) are the \code{log(shape1)} and
\code{log(shape2)}, respectively, for a beta distribution.
This distribution is used to generate quantiles to make the temperatures
not evenly distributed through the temperature space.}
}
\value{
A numeric vector for the temperatures to sample.
}
\description{
Make vector of temperatures for input of parameters
}
