% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TPCdesign-package.R
\docType{package}
\name{TPCdesign-package}
\alias{TPCdesign}
\alias{TPCdesign-package}
\title{TPCdesign: Design Informative Thermal Performance Experiments}
\description{
Tools to choose experimental temperatures that are maximally informative
for fitting thermal performance curves (TPCs). Given rough prior guesses
at a Brière-2 TPC's parameters, \code{\link[=design_temps]{design_temps()}} uses an approximate
coordinate-exchange (ACE) algorithm to search for the set of temperatures
that minimizes the expected uncertainty in fitted TPC parameters, and
\code{\link[=sim_gamma_data]{sim_gamma_data()}} simulates data from that curve for testing a design.
}
\section{Main functions}{

\itemize{
\item \code{\link[=design_temps]{design_temps()}}: Choose a set of temperatures to sample from.
\item \code{\link[=sim_gamma_data]{sim_gamma_data()}}: Simulate performance data with observation
error at a set of temperatures.
\item \code{\link[=briere2_tpc]{briere2_tpc()}}: Evaluate the Brière-2 TPC.
\item \code{\link[=briere2_tpc_Topt]{briere2_tpc_Topt()}}: Compute the thermal optimum of a Brière-2 TPC.
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/lucasnell/TPCdesign}
  \item Report bugs at \url{https://github.com/lucasnell/TPCdesign/issues}
}

}
\author{
\strong{Maintainer}: Lucas A. Nell \email{lucnell@gmail.com} (\href{https://orcid.org/0000-0003-3209-0517}{ORCID})

Authors:
\itemize{
  \item Lucas A. Nell \email{lucnell@gmail.com} (\href{https://orcid.org/0000-0003-3209-0517}{ORCID})
  \item Kelsey Lyberger \email{Kelsey.Lyberger@asu.edu} (\href{https://orcid.org/0000-0003-4903-380X}{ORCID})
}

}
\keyword{internal}
