% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{briere2_tpc}
\alias{briere2_tpc}
\title{Brière-2 thermal performance curve (TPC)}
\usage{
briere2_tpc(temp, CTmin, CTmax, a, b, scale = TRUE)
}
\arguments{
\item{temp}{Numeric vector of temperatures}

\item{CTmin}{Single numeric for parameter CTmin}

\item{CTmax}{Single numeric for parameter CTmax}

\item{a}{Single numeric for parameter a}

\item{b}{Single numeric for parameter b}

\item{scale}{Single logical for whether to scale to make max value 1.
Defaults to \code{TRUE}}
}
\value{
A numeric vector for measure of performance for each in \code{temp}
}
\description{
Brière-2 thermal performance curve (TPC)
}
