% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/designer.R
\name{design_temps}
\alias{design_temps}
\title{Design experimental temperatures}
\usage{
design_temps(
  n_temps,
  ctmin,
  ctmax,
  a,
  b,
  ctmin_err = 2.5,
  ctmax_err = 1.5,
  logb_err = 0.26,
  min_sep = 1,
  n_filler = 1L,
  n_draws = 250L,
  n_starts = 7L,
  digits = 2L
)
}
\arguments{
\item{n_temps}{Single integer for the number of temperatures to sample from.
Must be >= 2.}

\item{ctmin}{Single numeric for parameter \code{ctmin}.}

\item{ctmax}{Single numeric for parameter \code{ctmax}. Must be > \code{ctmin}.}

\item{a}{Single numeric for parameter \code{a}. Must be > 0.}

\item{b}{Single numeric for parameter \code{b}. Must be > 0.}

\item{ctmin_err}{Single numeric for error for \code{ctmin}.
This means priors for \code{ctmin} will be generated from a uniform
distribution with minimum \code{ctmin - ctmin_err} and
maximum \code{ctmin + ctmin_err}. Must be \verb{>= 0}.
Defaults to \code{2.50}.}

\item{ctmax_err}{Single numeric for error for \code{ctmax}.
This means priors for \code{ctmax} will be generated from a uniform
distribution with minimum \code{ctmax - ctmax_err} and
maximum \code{ctmax + ctmax_err}. Must be \verb{>= 0}.
Defaults to \code{1.50}.}

\item{logb_err}{Single numeric for error for \code{log(b)}.
This means priors for \code{log(b)} will be generated from a uniform
distribution with minimum \code{log(b) - logb_err} and
maximum \code{log(b) + logb_err}. Must be \verb{>= 0}.
Defaults to \code{0.26}.}

\item{min_sep}{Single numeric specifying the minimum separation between
optimized temperatures. The step where equally-spaced temperatures are
added (only happens when \code{n_filler >= 1}) is not affected by this
argument, and unless the number of points is very high relative
to the difference in maximum (\code{ctmax + ctmax_err}) and
minimum (\code{ctmin + ctmin_err}) temperatures surveyed, having a minimum
separation shouldn't change this step anyway. Defaults to \code{1}.}

\item{n_filler}{Single integer specifying the number of temperatures
that are equally spaced versus optimized.
Equally spacing points can be a good bet hedging strategy if you're
quite unsure of the TPC parameters (\code{ctmin}, \code{ctmax}, \code{a}, \code{b})
you're using. Defaults to \code{1L}.}

\item{n_draws}{Single integer specifying the number of prior draws to
use. Increase if your certainty is low (i.e., \verb{*_err} parameters are
high). Defaults to \code{250L}.}

\item{n_starts}{Single integer specifying the number of starts
Defaults to \code{7L}.}

\item{digits}{Single integer specifying the digits to round temperatures to.
Defaults to \code{2L}.}
}
\value{
A vector of temperatures at which to sample.
}
\description{
Design experimental temperatures
}
\examples{
design_temps(n_temps = 10, ctmin = 5, ctmax = 40, a = 1, b = 0.2)

# To run in parallel:
# library(future.apply)
# plan(multisession, workers = 4L)
# design_temps(n_temps = 10, ctmin = 5, ctmax = 40, a = 1, b = 0.2)

}
