% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting.R
\name{design_temps}
\alias{design_temps}
\title{Design experimental temperatures}
\usage{
design_temps(
  n_temps,
  n_reps,
  obs_cv,
  temp_min,
  temp_max,
  ctmin,
  ctmax,
  a,
  b,
  tpc_measure = "rmse",
  tpc_optim_fun = stats::optim,
  tpc_optim_args = list(),
  tpc_iter = 10L,
  tpc_lhstype = "improved",
  tpc_n_grid_temps = 101L,
  scale_tpc = TRUE,
  lower_bounds = NULL,
  upper_bounds = NULL,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{n_temps}{Single integer for the number of temperatures to sample from.}

\item{n_reps}{Single integer for the number of experimental reps per temperature.}

\item{obs_cv}{Single numeric for the observation error coefficient of variation.
Must be >= 0.
Observation error is simulated using a gamma distribution, except for
when performance values are below zero, in which case it uses
a normal approximation.}

\item{temp_min}{Single numeric for the minimum possible temperature allowed.}

\item{temp_max}{Single numeric for the maximum possible temperature allowed.}

\item{ctmin}{Single numeric for parameter \code{ctmin}.}

\item{ctmax}{Single numeric for parameter \code{ctmax}.}

\item{a}{Single numeric for parameter \code{a}.}

\item{b}{Single numeric for parameter \code{b}.}

\item{tpc_measure}{Single string for the measure to use to fit temperatures.
Options are \code{"rmse"}, \code{"ctmin"}, \code{"ctmax"}, \code{"b"}, or \code{"a"}, and
capitalization is ignored.
Option \code{"rmse"} returns the root mean square deviation between the
between "observed" (simulated) and expected performance values for a
given set of parameters.
The other options are absolute differences between true and fitted
values for individual TPC parameter values.
Defaults to \code{"rmse"}.}

\item{tpc_optim_fun}{A function to use for optimization in TPC fitting.
Options are \code{stats::optim} or functions from the \code{nloptr} package.
Defaults to \code{stats::optim}.}

\item{tpc_optim_args}{A named list of additional arguments to pass
to \code{tpc_optim_fun}.
Defaults \code{list()}.}

\item{tpc_iter}{Single integer specifying the
"number of partitions (simulations or design points or rows)"
to use in Latin Hypercube Sampling for TPC fitting.}

\item{tpc_lhstype}{Single string specifying the type of Latin Hypercube
Sampling to use (via the \code{lhs} package) for TPC fitting. Options are
\code{"random"} (uses \link[lhs:randomLHS]{lhs::randomLHS})
\code{"improved"} (uses \link[lhs:improvedLHS]{lhs::improvedLHS})
\code{"maximin"} (uses \link[lhs:maximinLHS]{lhs::maximinLHS})
\code{"genetic"} (uses \link[lhs:geneticLHS]{lhs::geneticLHS}).
Defaults to \code{"improved"}.}

\item{tpc_n_grid_temps}{Single interger indicating the number of
temperatures to use to compute RMSE in TPC fitting. Must be \verb{>= 10}.
Defaults to \code{101L}.}

\item{scale_tpc}{Single logical for whether to scale performance values
to make them have a max value of 1.
Defaults to \code{TRUE}.}

\item{lower_bounds}{Numeric vectors of length \code{n_temps+1L} indicating
the lower bounds in starting conditions used in
\code{estimatePMR::winnowing_optim}.
Values must be > 0 and < 1.
If not provided, it defaults to \code{rep(1e-9, n_temps+1L)}.}

\item{upper_bounds}{Numeric vectors of length \code{n_temps+1L} indicating
the upper bounds in starting conditions used in
\code{estimatePMR::winnowing_optim}.
Values must be > 0 and < 1.
If not provided, it defaults to \code{rep(1 - 1e-9, n_temps+1L)}.}

\item{verbose}{Single logical for whether to warn of poor fits.
Useful for debugging, but results in potentially many warnings.
Defaults to \code{FALSE}.}

\item{...}{Other arguments to pass to \code{\link[estimatePMR:winnowing_optim]{estimatePMR::winnowing_optim()}}.
Arguments \code{fn} and \code{fn_args} are not allowed to be passed here because
they are handled by this function explicitly.
If nothing is provided here, \code{winnowing_optim} is run with
\code{n_bevals = 100L},
\code{n_boxes = 10L},
\code{n_outputs = 1L},
\code{controls = list(list(maxit = 1000, reltol = 1e-08))},
and
\code{optimizers = c(nloptr::newuoa)}.}
}
\value{
Typically the output will be a single object of the type returned
by \code{optim_fun}. The exception is if the \code{n_outputs} is provided
(and passed to \code{winnowing_optim()}), and the last item in \code{n_outputs}
is greater than 1.
In that case, it will be a list of items returned from \code{optim_fun},
and the list's length will be \code{tail(n_outputs, 1L)}.
Regardless of length, each item in the output has an extra field added
to it (\code{"temps"}) indicating the temperatures from this fit.
}
\description{
Design experimental temperatures
}
