% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{make_temps}
\alias{make_temps}
\title{Make vector of temperatures for input of parameters}
\usage{
make_temps(params, temp_min, temp_max)
}
\arguments{
\item{params}{Numeric vector of parameters.
The length of this vector must be equal to the number of desired
temperatures plus 1. A vector of length \verb{<2} returns an error.
The last item in this vector must be the logit-transformed proportion
the last temperature is from the midpoint
(i.e., \code{mean(c(temp_min, temp_max))})
to the maximum possible temperature (i.e., \code{temp_max}).
The first item is the logit-transformed weight affecting
the difference between \code{temp_min} and the first sampled temperature.
The remaining items are the logit-transformed weights affecting
the distance between the remaining sampled temperatures and
the previous ones.}

\item{temp_min}{Single numeric for the minimum possible temperature allowed.}

\item{temp_max}{Single numeric for the maximum possible temperature allowed.}
}
\value{
A numeric vector for the temperatures to sample.
}
\description{
Make vector of temperatures for input of parameters
}
